pm: Move meta codegen from build.rs
