+++ pm/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "lox"
-version = "0.0.0"
+name = "pm"
+version = "0.1.0"
 edition = "2021"
+
+[lib]
+proc-macro = true
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fs-err = "2.7.0"
-phf = { version = "0.10.1", features = ["macros"] }
-pm = { path = './pm' }
